fix(sdk): publish generated Ordex 1.2 modules

+++ sdk/src/swaps.ts
@@ -441,17 +441,18 @@ export function verifySwapAcceptance(acceptance: unknown, intent: unknown): Swap
   if (fee !== declaredFee) {
     return acceptanceRefuse('FEE_CHANGED', 'The declared fee does not match the transaction.');
   }
-  if (makerFee > parseSats(i.maxMakerFeeSats)) {
+  if (makerFee > (parseSats(i.maxMakerFeeSats) as bigint)) {
     return acceptanceRefuse('FEE_BUDGET_EXCEEDED', 'The maker contribution exceeds the budget the intent approved.');
   }
 
   // Each required criterion must be satisfied by an output paying the maker
   // receive script with at least the minimum quantity.
   for (const criterion of i.requires as SwapRequirement[]) {
-    const minimum = parseSats(criterion.minQuantitySats);
+    const minimum = parseSats(criterion.minQuantitySats) as bigint;
     const satisfied = outputs.some(
       (output) =>
-        output.scriptHex === i.makerReceiveScriptHex && parseSats(output?.valueSats) >= minimum,
+        output.scriptHex === i.makerReceiveScriptHex &&
+        (parseSats(output?.valueSats) as bigint) >= minimum,
     );
     if (!satisfied) {
       return acceptanceRefuse(
